dist: Add 'put -t <type>' and fix 'insert' -> 'put'
If we want to get the object type out with 'get -t <id>', we'll have
to either set it on 'put' or auto-detect it.  I'd rather set it
explicitly to avoid having to code auto-detection logic.

Besides layers, we'll likely also want to put the signed portion of
the manifest (or individual history nuggets?).  In any case, we'll
need more than layer tarballs to build the manifest.